mv server event func -> process_event.cpp
